+++ src/NEventStore.Persistence.MongoDB.Tests/AcceptanceTests/Issues/Issue8.cs
@@ -18,54 +18,54 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests.Issues
 {
 #if MSTEST
-    [TestClass]
+	[TestClass]
 #endif
-    public class Issue8 : SpecificationBase
-    {
-        private Exception _error;
-        private const string InvalidConnectionStringName = "this_is_not_a_connection_string";
+	public class Issue8 : SpecificationBase
+	{
+		private Exception? _error;
+		private const string InvalidConnectionStringName = "this_is_not_a_connection_string";
 
-        protected override void Context()
-        {
-        }
+		protected override void Context()
+		{
+		}
 
-        protected override void Because()
-        {
-            _error = Catch.Exception(() =>
-            {
-                Wireup.Init()
-                    .UsingMongoPersistence(InvalidConnectionStringName, new DocumentObjectSerializer())
-                    .Build();
-            });
-        }
+		protected override void Because()
+		{
+			_error = Catch.Exception(() =>
+			{
+				Wireup.Init()
+					.UsingMongoPersistence(InvalidConnectionStringName, new DocumentObjectSerializer())
+					.Build();
+			});
+		}
 
-#if NET462
-        /// <summary>
-        /// in net462, UsingMongoPersistence will look for a connection string name in the app.config
-        /// and it will not find anything
-        /// </summary>
-        [Fact]
-        public void A_configuration_exception_should_be_thrown()
-        {
-            _error.Should().BeOfType<NEventStore.Persistence.MongoDB.ConfigurationException>();
-        }
-#endif
-
-#if !NET462
-        /// <summary>
-        /// in netstandard2.0, UsingMongoPersistence will accept a connectionString which will be invalid
-        /// </summary>
-        [Fact]
-        public void A_configuration_exception_should_be_thrown()
-        {
-            _error.Should().BeOfType<MongoConfigurationException>();
-        }
+#if NET472_OR_GREATER
+		/// <summary>
+		/// in net462, UsingMongoPersistence will look for a connection string name in the app.config
+		/// and it will not find anything
+		/// </summary>
+		[Fact]
+		public void A_configuration_exception_should_be_thrown()
+		{
+			_error!.Should().BeOfType<NEventStore.Persistence.MongoDB.ConfigurationException>();
+		}
 #endif
 
-        [Fact]
-        public void A_configuration_error_should_be_thrown()
-        {
-            _error.Message.Should().Contain(InvalidConnectionStringName);
-        }
-    }
+#if !NET472_OR_GREATER
+		/// <summary>
+		/// in netstandard2.0, UsingMongoPersistence will accept a connectionString which will be invalid
+		/// </summary>
+		[Fact]
+		public void A_configuration_exception_should_be_thrown()
+		{
+			_error!.Should().BeOfType<MongoConfigurationException>();
+		}
+#endif
+
+		[Fact]
+		public void A_configuration_error_should_be_thrown()
+		{
+			_error!.Message.Should().Contain(InvalidConnectionStringName);
+		}
+	}
 }