+++ src/NEventStore.Persistence.MongoDB/ExtensionMethods.cs
@@ -1,38 +1,23 @@
-﻿using MongoDB.Bson.Serialization;
-using MongoDB.Bson.Serialization.Attributes;
+﻿using MongoDB.Bson.Serialization.Attributes;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using global::MongoDB.Bson;
+using global::MongoDB.Bson.Serialization.Options;
+using global::MongoDB.Driver;
+using NEventStore.Serialization;
+using BsonSerializer = global::MongoDB.Bson.Serialization.BsonSerializer;
 
 namespace NEventStore.Persistence.MongoDB
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using global::MongoDB.Bson;
-    using global::MongoDB.Bson.IO;
-    using global::MongoDB.Bson.Serialization.Options;
-    using global::MongoDB.Bson.Serialization.Serializers;
-    using global::MongoDB.Driver;
-    using NEventStore.Serialization;
-    using BsonSerializer = global::MongoDB.Bson.Serialization.BsonSerializer;
-
+    /// <summary>
+    /// Extension methods for working with MongoDB.
+    /// </summary>
     public static class ExtensionMethods
     {
-        [Obsolete("Original code, not used anymore, replaced by the new configurable version")]
-        public static Dictionary<Tkey, Tvalue> AsDictionary<Tkey, Tvalue>(this BsonValue bsonValue)
-        {
-            using (BsonReader reader = new JsonReader(bsonValue.ToJson()))
-            {
-                var dictionarySerializer = new DictionaryInterfaceImplementerSerializer<Dictionary<Tkey, Tvalue>>();
-                object result = dictionarySerializer.Deserialize(
-                    BsonDeserializationContext.CreateRoot(reader, _ => { }),
-                    new BsonDeserializationArgs()
-                    {
-                        NominalType = typeof(Dictionary<Tkey, Tvalue>)
-                    }
-                );
-                return (Dictionary<Tkey, Tvalue>)result;
-            }
-        }
-
+        /// <summary>
+        /// Converts a <see cref="CommitAttempt"/> to a <see cref="BsonDocument"/>.
+        /// </summary>
         public static BsonDocument ToMongoCommit(this CommitAttempt commit, Int64 checkpoint, IDocumentSerializer serializer)
         {
             int streamRevision = commit.StreamRevision - (commit.Events.Count - 1);
@@ -64,6 +49,10 @@ namespace NEventStore.Persistence.MongoDB
             return mc.ToBsonDocument();
         }
 
+        /// <summary>
+        /// Converts a <see cref="CommitAttempt"/> to a <see cref="BsonDocument"/> representing an empty commit.
+        /// </summary>
+        /// <exception cref="ArgumentNullException"></exception>
         public static BsonDocument ToEmptyCommit(this CommitAttempt commit, Int64 checkpoint, String systemBucketName)
         {
             if (commit == null) throw new ArgumentNullException(nameof(commit));
@@ -87,38 +76,9 @@ namespace NEventStore.Persistence.MongoDB
             return mc.ToBsonDocument();
         }
 
-        [Obsolete("Original code, not used anymore, replaced by the new configurable version")]
-        public static BsonDocument ToMongoCommit_original(this CommitAttempt commit, Int64 checkpoint, IDocumentSerializer serializer)
-        {
-            int streamRevision = commit.StreamRevision - (commit.Events.Count - 1);
-            int streamRevisionStart = streamRevision;
-            IEnumerable<BsonDocument> events = commit
-                .Events
-                .Select(e =>
-                    new BsonDocument
-                    {
-                        {MongoCommitFields.StreamRevision, streamRevision++},
-                        {MongoCommitFields.Payload, BsonDocumentWrapper.Create(typeof(EventMessage), serializer.Serialize(e))}
-                    });
-
-            //var dictionarySerialize = new DictionaryInterfaceImplementerSerializer<Dictionary<string, object>>(DictionaryRepresentation.ArrayOfArrays);
-            //var dicSer = BsonSerializer.LookupSerializer<Dictionary<string, object>>();
-
-            return new BsonDocument
-            {
-                {MongoCommitFields.CheckpointNumber, checkpoint},
-                {MongoCommitFields.CommitId, commit.CommitId},
-                {MongoCommitFields.CommitStamp, commit.CommitStamp},
-                {MongoCommitFields.Headers, BsonDocumentWrapper.Create(commit.Headers)},
-                {MongoCommitFields.Events, new BsonArray(events)},
-                {MongoCommitFields.StreamRevisionFrom, streamRevisionStart},
-                {MongoCommitFields.StreamRevisionTo, streamRevision - 1},
-                {MongoCommitFields.BucketId, commit.BucketId},
-                {MongoCommitFields.StreamId, commit.StreamId},
-                {MongoCommitFields.CommitSequence, commit.CommitSequence}
-            };
-        }
-
+        /// <summary>
+        /// Converts a <see cref="BsonDocument"/> to a <see cref="ICommit"/>.
+        /// </summary>
         public static ICommit ToCommit(this BsonDocument doc, IDocumentSerializer serializer)
         {
             if (doc == null)
@@ -145,57 +105,27 @@ namespace NEventStore.Persistence.MongoDB
                 }).ToArray());
         }
 
-        [Obsolete("Original code, not used anymore, replaced by the new configurable version")]
-        public static ICommit ToCommit_original(this BsonDocument doc, IDocumentSerializer serializer)
-        {
-            if (doc == null)
-            {
-                return null;
-            }
-
-            string bucketId = doc[MongoCommitFields.BucketId].AsString;
-            string streamId = doc[MongoCommitFields.StreamId].AsString;
-            int commitSequence = doc[MongoCommitFields.CommitSequence].AsInt32;
-
-            var events = doc[MongoCommitFields.Events]
-                .AsBsonArray
-                .Select(e =>
-                {
-                    BsonValue payload = e.AsBsonDocument[MongoCommitFields.Payload];
-                    return payload.IsBsonDocument
-                           ? BsonSerializer.Deserialize<EventMessage>(payload.ToBsonDocument())
-                           : serializer.Deserialize<EventMessage>(payload.AsByteArray);
-                })
-                .ToArray();
-
-            //int streamRevision = doc[MongoCommitFields.Events].AsBsonArray.Last().AsBsonDocument[MongoCommitFields.StreamRevision].AsInt32;
-            int streamRevision = doc[MongoCommitFields.StreamRevisionTo].AsInt32;
-            return new Commit(bucketId,
-                streamId,
-                streamRevision,
-                doc[MongoCommitFields.CommitId].AsGuid,
-                commitSequence,
-                doc[MongoCommitFields.CommitStamp].ToUniversalTime(),
-                doc[MongoCommitFields.CheckpointNumber].ToInt64(),
-                doc[MongoCommitFields.Headers].AsDictionary<string, object>(),
-                events);
-        }
-
+        /// <summary>
+        /// Converts a <see cref="ISnapshot"/> to a <see cref="BsonDocument"/>.
+        /// </summary>
         public static BsonDocument ToMongoSnapshot(this ISnapshot snapshot, IDocumentSerializer serializer)
         {
             return new BsonDocument
             {
-                { MongoShapshotFields.Id, new BsonDocument
+                { MongoSnapshotFields.Id, new BsonDocument
                     {
-                        {MongoShapshotFields.BucketId, snapshot.BucketId},
-                        {MongoShapshotFields.StreamId, snapshot.StreamId},
-                        {MongoShapshotFields.StreamRevision, snapshot.StreamRevision}
+                        {MongoSnapshotFields.BucketId, snapshot.BucketId},
+                        {MongoSnapshotFields.StreamId, snapshot.StreamId},
+                        {MongoSnapshotFields.StreamRevision, snapshot.StreamRevision}
                     }
                 },
-                { MongoShapshotFields.Payload, BsonDocumentWrapper.Create(serializer.Serialize(snapshot.Payload)) }
+                { MongoSnapshotFields.Payload, BsonDocumentWrapper.Create(serializer.Serialize(snapshot.Payload)) }
             };
         }
 
+        /// <summary>
+        /// Converts a <see cref="BsonDocument"/> to a <see cref="ISnapshot"/>.
+        /// </summary>
         public static Snapshot ToSnapshot(this BsonDocument doc, IDocumentSerializer serializer)
         {
             if (doc == null)
@@ -203,11 +133,11 @@ namespace NEventStore.Persistence.MongoDB
                 return null;
             }
 
-            BsonDocument id = doc[MongoShapshotFields.Id].AsBsonDocument;
-            string bucketId = id[MongoShapshotFields.BucketId].AsString;
-            string streamId = id[MongoShapshotFields.StreamId].AsString;
-            int streamRevision = id[MongoShapshotFields.StreamRevision].AsInt32;
-            BsonValue bsonPayload = doc[MongoShapshotFields.Payload];
+            BsonDocument id = doc[MongoSnapshotFields.Id].AsBsonDocument;
+            string bucketId = id[MongoSnapshotFields.BucketId].AsString;
+            string streamId = id[MongoSnapshotFields.StreamId].AsString;
+            int streamRevision = id[MongoSnapshotFields.StreamRevision].AsInt32;
+            BsonValue bsonPayload = doc[MongoSnapshotFields.Payload];
 
             object payload;
             switch (bsonPayload.BsonType)
@@ -226,6 +156,9 @@ namespace NEventStore.Persistence.MongoDB
             return new Snapshot(bucketId, streamId, streamRevision, payload);
         }
 
+        /// <summary>
+        /// Converts a <see cref="BsonDocument"/> to a <see cref="StreamHead"/>.
+        /// </summary>
         public static StreamHead ToStreamHead(this BsonDocument doc)
         {
             BsonDocument id = doc[MongoStreamHeadFields.Id].AsBsonDocument;
@@ -234,6 +167,9 @@ namespace NEventStore.Persistence.MongoDB
             return new StreamHead(bucketId, streamId, doc[MongoStreamHeadFields.HeadRevision].AsInt32, doc[MongoStreamHeadFields.SnapshotRevision].AsInt32);
         }
 
+        /// <summary>
+        /// Returns a query that can be used to find a commit given the information of the <see cref="CommitAttempt"/>.
+        /// </summary>
         public static FilterDefinition<BsonDocument> ToMongoCommitIdQuery(this CommitAttempt commit)
         {
             var builder = Builders<BsonDocument>.Filter;
@@ -244,62 +180,93 @@ namespace NEventStore.Persistence.MongoDB
             );
         }
 
-        public static FilterDefinition<BsonDocument> ToMongoCommitIdQuery(this ICommit commit)
-        {
-            var builder = Builders<BsonDocument>.Filter;
-            return builder.And(
-                builder.Eq(MongoCommitFields.BucketId, commit.BucketId),
-                builder.Eq(MongoCommitFields.StreamId, commit.StreamId),
-                builder.Eq(MongoCommitFields.CommitId, commit.CommitId)
-            );
-        }
-
+        /// <summary>
+        /// Builds a query that can be used to find the snapshot for a stream.
+        /// </summary>
         public static FilterDefinition<BsonDocument> GetSnapshotQuery(string bucketId, string streamId, int maxRevision)
         {
             var builder = Builders<BsonDocument>.Filter;
 
             return builder.And(
-                builder.Eq(MongoShapshotFields.FullQualifiedBucketId, bucketId),
-                builder.Eq(MongoShapshotFields.FullQualifiedStreamId, streamId),
-                builder.Lte(MongoShapshotFields.FullQualifiedStreamRevision, maxRevision)
+                builder.Eq(MongoSnapshotFields.FullQualifiedBucketId, bucketId),
+                builder.Eq(MongoSnapshotFields.FullQualifiedStreamId, streamId),
+                builder.Lte(MongoSnapshotFields.FullQualifiedStreamRevision, maxRevision)
             );
         }
     }
 
-    // let's ignore the extra elements, the 'Dispatched' field and the dispatched concept have been dropped in NEventStore 6
+    /// <summary>
+    /// Represents a commit in MongoDB.
+    /// </summary>
+    /// <remarks>
+    /// let's ignore the extra elements, the 'Dispatched' field and the dispatched concept have been dropped in NEventStore 6
+    /// </remarks>
     [BsonIgnoreExtraElements]
     public sealed class MongoCommit
     {
+        /// <summary>
+        /// Gets or sets the checkpoint number.
+        /// </summary>
         [BsonId]
         [BsonElement(MongoCommitFields.CheckpointNumber)]
         public long CheckpointNumber { get; set; }
 
+        /// <summary>
+        /// Gets or sets the commit identifier.
+        /// </summary>
         [BsonElement(MongoCommitFields.CommitId)]
         public Guid CommitId { get; set; }
 
+        /// <summary>
+        /// Gets or sets the commit timestamp.
+        /// </summary>
         [BsonElement(MongoCommitFields.CommitStamp)]
         public DateTime CommitStamp { get; set; }
 
+        /// <summary>
+        /// Gets or sets the commit headers.
+        /// </summary>
+        /// <remarks>
+        /// we can override this specifying a ClassMap OR implementing an <see cref="global::MongoDB.Bson.Serialization.IBsonSerializer"/>
+        /// and an <see cref="global::MongoDB.Bson.Serialization.IBsonSerializationProvider"/>
+        /// </remarks>
         [BsonElement(MongoCommitFields.Headers)]
-        [BsonDictionaryOptions(DictionaryRepresentation.ArrayOfArrays)] // we can override this specifing a classmap OR implementing an IBsonSerializer and an IBsonSerializationProvider 
+        [BsonDictionaryOptions(DictionaryRepresentation.ArrayOfArrays)]
         public IDictionary<string, object> Headers { get; set; }
 
-        // multiple evaluations using linq can be dangerous, maybe it's better have a plain array to avoid bugs
+        /// <summary>
+        /// Gets or sets the events.
+        /// </summary>
         [BsonElement(MongoCommitFields.Events)]
-        public BsonArray Events { get; set; }
+        public BsonArray Events { get; set; } // multiple evaluations using linq can be dangerous, maybe it's better have a plain array to avoid bugs
 
+        /// <summary>
+        /// Gets or sets the stream revision from.
+        /// </summary>
         [BsonElement(MongoCommitFields.StreamRevisionFrom)]
         public int StreamRevisionFrom { get; set; }
 
+        /// <summary>
+        /// Gets or sets the stream revision to.
+        /// </summary>
         [BsonElement(MongoCommitFields.StreamRevisionTo)]
         public int StreamRevisionTo { get; set; }
 
+        /// <summary>
+        /// Gets or sets the bucket identifier.
+        /// </summary>
         [BsonElement(MongoCommitFields.BucketId)]
         public string BucketId { get; set; }
 
+        /// <summary>
+        /// Gets or sets the stream identifier.
+        /// </summary>
         [BsonElement(MongoCommitFields.StreamId)]
         public string StreamId { get; set; }
 
+        /// <summary>
+        /// Gets or sets the commit sequence.
+        /// </summary>
         [BsonElement(MongoCommitFields.CommitSequence)]
         public int CommitSequence { get; set; }
     }